Operation name conflict resolving in operation name generators.
