+++ src/NSwag.CodeGeneration/ClientGeneratorBase.cs
@@ -109,7 +109,7 @@ namespace NSwag.CodeGeneration
         {
             document.GenerateOperationIds();
 
-            var operations = document.Paths
+            return document.Paths
                 .SelectMany(pair => pair.Value.Select(p => new { Path = pair.Key.Trim('/'), HttpMethod = p.Key, Operation = p.Value }))
                 .Select(tuple =>
                 {
@@ -121,166 +121,6 @@ namespace NSwag.CodeGeneration
                     return operationModel;
                 })
                 .ToList();
-
-            operations = ResolveOperationConflicts(document, operations);
-
-            return operations;
-        }
-
-        /// <summary>
-        /// Operation model comparer.
-        /// </summary>
-        internal class OperationComparer : IEqualityComparer<TOperationModel>
-        {
-            public bool Equals(TOperationModel x, TOperationModel y)
-            {
-                // Check whether the compared objects reference the same data.
-                if (ReferenceEquals(x, y)) return true;
-
-                // Check whether any of the compared objects is null.
-                if (ReferenceEquals(x, null) || ReferenceEquals(y, null))
-                    return false;
-
-                // Check whether the operations' properties are equal.
-                return (x.ControllerName == y.ControllerName)
-                    && (x.OperationName == y.OperationName)
-                    && (x.Parameters.SequenceEqual(y.Parameters, new OperationParameterComparer()));
-            }
-
-            public int GetHashCode(TOperationModel operation)
-            {
-                // Check whether the object is null
-                if (ReferenceEquals(operation, null)) return 0;
-
-                // Calculate the hash code.
-                var hash = string.IsNullOrEmpty(operation.ControllerName) ? 0 : operation.ControllerName.GetHashCode();
-                hash ^= string.IsNullOrEmpty(operation.OperationName) ? 0 : operation.OperationName.GetHashCode();
-                foreach (var parameter in operation.Parameters)
-                {
-                    hash ^= string.IsNullOrEmpty(parameter.Type) ? 0 : parameter.Type.GetHashCode();
-                }
-                return hash;
-            }
-        }
-
-        /// <summary>
-        /// Operation parameter comparer.
-        /// </summary>
-        internal class OperationParameterComparer : IEqualityComparer<TParameterModel>
-        {
-            public bool Equals(TParameterModel x, TParameterModel y)
-            {
-                // Check whether the compared objects reference the same data.
-                if (ReferenceEquals(x, y)) return true;
-
-                // Check whether any of the compared objects is null.
-                if (ReferenceEquals(x, null) || ReferenceEquals(y, null))
-                    return false;
-
-                // Check whether the parameters' properties are equal.
-                return x.Type == y.Type;
-            }
-
-            public int GetHashCode(TParameterModel parameter)
-            {
-                // Check whether the object is null
-                if (ReferenceEquals(parameter, null)) return 0;
-
-                // Get hash code for the Type field if it is not null.
-                return string.IsNullOrEmpty(parameter.Type) ? 0 : parameter.Type.GetHashCode();
-            }
-        }
-
-        /// <summary>
-        /// Enumeration of operation name deconfliction suffixes.
-        /// </summary>
-        internal enum DeconflictSuffix
-        {
-            /// <summary>
-            /// Use underlying HTTP method as name suffix.
-            /// </summary>
-            HttpMethod,
-
-            /// <summary>
-            /// Use operation index as name suffix.
-            /// </summary>
-            OperationIndex,
-
-            /// <summary>
-            /// Placeholder for enumeration end
-            /// </summary>
-            End
-        }
-
-        /// <summary>
-        /// Tries to resolve controller operations with overlapping names and signatures.
-        /// </summary>
-        /// <param name="document">Original Swagger document.</param>
-        /// <param name="operations">Operations extracted from Swagger document.</param>
-        /// <returns>Deduplicated operations.</returns>
-        internal static List<TOperationModel> ResolveOperationConflicts(SwaggerDocument document, List<TOperationModel> operations)
-        {
-            var ret = operations;
-            if (ret == null || ret.Count == 0)
-            {
-                return ret;
-            }
-            // Set initial deconfliction suffix
-            var suffix = DeconflictSuffix.HttpMethod;
-            List<IGrouping<TOperationModel, TOperationModel>> conflicts = null;
-            do
-            {
-                conflicts = Deconflict(ref operations, conflicts, suffix++);
-            }
-            while (conflicts.Any() && (suffix < DeconflictSuffix.End));
-
-            return ret;
-        }
-
-        internal static List<IGrouping<TOperationModel, TOperationModel>> Deconflict(ref List<TOperationModel> operations, List<IGrouping<TOperationModel, TOperationModel>> conflicts, DeconflictSuffix suffix)
-        {
-            var comparer = new OperationComparer();
-            if (conflicts == null)
-            {
-                conflicts = operations.GroupBy(x => x, comparer).Where(g => g.Count() > 1).ToList();
-            }
-            if (conflicts.Any())
-            {
-                foreach (var conflict in conflicts)
-                {
-                    for (var idx = 0; idx < conflict.Count(); idx++)
-                    {
-                        var operation = conflict.ElementAt(idx);
-                        switch (suffix)
-                        {
-                            case DeconflictSuffix.HttpMethod:
-                                operation.OperationName += CapitalizeFirst(operation.HttpMethod.ToString());
-                                break;
-
-                            case DeconflictSuffix.OperationIndex:
-                                operation.OperationName += $"{idx + 1}";
-                                break;
-
-                            default:
-                                break;
-                        }
-                    }
-                }
-            }
-            return operations.GroupBy(x => x, comparer).Where(g => g.Count() > 1).ToList();
-        }
-
-        /// <summary>Capitalizes first letter.</summary>
-        /// <param name="name">The name to capitalize.</param>
-        /// <returns>Capitalized name.</returns>
-        internal static string CapitalizeFirst(string name)
-        {
-            if (string.IsNullOrEmpty(name))
-            {
-                return string.Empty;
-            }
-            var capitalized = name.ToLower();
-            return char.ToUpper(capitalized[0]) + (capitalized.Length > 1 ? capitalized.Substring(1) : "");
         }
     }
 }
