+++ src/NSwag.CodeGeneration/OperationNameGenerators/MultipleClientsFromPathSegmentsOperationNameGenerator.cs
@@ -40,11 +40,47 @@ namespace NSwag.CodeGeneration.OperationNameGenerators
         /// <returns>The client name.</returns>
         public string GetOperationName(SwaggerDocument document, string path, SwaggerOperationMethod httpMethod, SwaggerOperation operation)
         {
+            var operationName = PathToName(path);
+            var nameConflict = document.Paths
+                                   .SelectMany(pair => pair.Value.Select(p => new { Path = pair.Key.Trim('/'), HttpMethod = p.Key, Operation = p.Value }))
+                                   .Where(op =>
+                                           GetClientName(document, op.Path, op.HttpMethod, op.Operation) == GetClientName(document, path, httpMethod, operation)
+                                           && PathToName(op.Path) == operationName
+                                       //&& op.Operation.Parameters.Count == operation.Parameters.Count // Compare by operation signature
+                                   )
+                                   .ToList().Count > 1;
+            if (nameConflict)
+            {
+                operationName += CapitalizeFirst(httpMethod.ToString());
+            }
+            return operationName;
+        }
+
+        /// <summary>
+        /// Converts the path to an operation name.
+        /// </summary>
+        /// <param name="path">The HTTP path.</param>
+        /// <returns>The operation name.</returns>
+        internal static string PathToName(string path)
+        {
             return path
                 .Split('/')
                 .Where(p => !p.Contains("{") && !string.IsNullOrWhiteSpace(p))
                 .Reverse()
                 .FirstOrDefault() ?? "Index";
+        }
+
+        /// <summary>Capitalizes first letter.</summary>
+        /// <param name="name">The name to capitalize.</param>
+        /// <returns>Capitalized name.</returns>
+        internal static string CapitalizeFirst(string name)
+        {
+            if (string.IsNullOrEmpty(name))
+            {
+                return string.Empty;
+            }
+            var capitalized = name.ToLower();
+            return char.ToUpper(capitalized[0]) + (capitalized.Length > 1 ? capitalized.Substring(1) : "");
         }
     }
 }
